[MOV] sale_project_timesheet: extract reinvoicing timesheet from project flow
Impacted modules: sale_timesheet and sale_project_timesheet

This commit factorize the billing mecanism of timesheet to be reuse
later by other modules. We now have :
- sale_timesheet: billing mecanism by linking sale lines to timesheet
line, and increase delivered quantity. It also provide the basics for
service-product configuration and invoicing type of timesheet for
reporting.
- sale_project_timesheet: determines which sale line should be applied
on timesheet linked to project/tasks. It also provides reporting tool
per project to analyse costs.

Nothing should change functionnally, as this commit only move code
from an existing module to another new one.

Task-1923949
