+++ addons/sale_project_timesheet/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import account
-from . import account_invoice
-from . import product
-from . import sale_order
+from . import models
+from . import wizard
+from . import report
