+++ addons/sale_project_timesheet/models/__init__.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import account
-from . import account_invoice
 from . import product
+from . import project
+from . import project_overview
+from . import project_sale_line_employee_map
 from . import sale_order
+from . import timesheet
