+++ addons/sale_project_timesheet/static/src/scss/sale_timesheet.scss
