+++ addons/sale_project_timesheet/tests/test_reinvoice.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.sale_timesheet.tests.common import TestCommonSaleTimesheetNoChart
+from odoo.addons.sale_project_timesheet.tests.common import TestCommonSaleTimesheetNoChart
 
 
 class TestReInvoice(TestCommonSaleTimesheetNoChart):
