+++ addons/sale_project_timesheet/views/hr_timesheet_templates.xml
@@ -3,7 +3,7 @@
 
     <template id="assets_backend" name="sale timesheet assets" inherit_id="web.assets_backend">
         <xpath expr="." position="inside">
-            <link rel="stylesheet" type="text/scss" href="/sale_timesheet/static/src/scss/sale_timesheet.scss"/>
+            <link rel="stylesheet" type="text/scss" href="/sale_project_timesheet/static/src/scss/sale_timesheet.scss"/>
         </xpath>
     </template>
