+++ addons/sale_project_timesheet/views/product_views.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="product_template_view_form" model="ir.ui.view">
+        <field name="name">product.template.timesheet.form</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="sale_timesheet.view_product_timesheet_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='service_tracking']" position="after">
+                <field name="project_id" attrs="{'invisible':[('service_tracking','!=','task_global_project')], 'required':[('service_tracking','==','task_global_project')]}"/>
+                <field name="project_template_id" context="{'active_test': False}" attrs="{'invisible':[('service_tracking','not in',['task_new_project', 'project_only'])]}"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
