+++ addons/sale_project_timesheet/views/project_overview_views.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <!--
+        Plan
+    -->
+    <record id="timesheet_action_plan_pivot" model="ir.actions.act_window">
+        <field name="name">Timesheet</field>
+        <field name="res_model">account.analytic.line</field>
+        <field name="view_mode">pivot,tree,form</field>
+        <field name="domain">[('is_timesheet', '=', True)]</field>
+        <field name="search_view_id" ref="hr_timesheet.hr_timesheet_line_search"/>
+    </record>
+
+    <record id="timesheet_action_from_plan" model="ir.actions.act_window">
+        <field name="name">Timesheet</field>
+        <field name="res_model">account.analytic.line</field>
+        <field name="view_mode">tree,form</field>
+        <field name="domain">[('is_timesheet', '=', True)]</field>
+        <field name="search_view_id" ref="hr_timesheet.hr_timesheet_line_search"/>
+    </record>
+
+</odoo>
