+++ addons/sale_project_timesheet/views/sale_order_views.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+        <record id="sale_order_view_form" model="ir.ui.view">
+            <field name="name">sale.order.form.inherit.sale.project.timesheet</field>
+            <field name="model">sale.order</field>
+            <field name="inherit_id" ref="sale.view_order_form"/>
+            <field name="arch" type="xml">
+                <data>
+                    <xpath expr="//button[@name='action_view_invoice']" position="before">
+                       <button type="object"
+                           name="action_view_project_ids"
+                           class="oe_stat_button"
+                           icon="fa-puzzle-piece"
+                           string="Project(s) Overview"
+                           attrs="{'invisible': [('project_ids', '=', [])]}"
+                           groups="project.group_project_manager">
+                          <field name="project_ids" invisible="1"/>
+                       </button>
+                    </xpath>
+                    <xpath expr="//button[@name='action_view_invoice']" position="before">
+                       <button type="object"
+                           name="action_view_task"
+                           class="oe_stat_button"
+                           icon="fa-tasks"
+                           attrs="{'invisible': [('tasks_count', '=', 0)]}"
+                           groups="project.group_project_user">
+                           <field name="tasks_count" widget="statinfo" string="Tasks"/>
+                       </button>
+                    </xpath>
+                </data>
+           </field>
+        </record>
+</odoo>
