+++ addons/sale_timesheet/__init__.py
@@ -4,8 +4,7 @@
 from odoo import api, SUPERUSER_ID
 
 from . import models
-from . import wizard
-from . import report
+
 
 def uninstall_hook(cr, registry):
     env = api.Environment(cr, SUPERUSER_ID, {})
