+++ odoo/addons/test_main_flows/__manifest__.py
@@ -8,7 +8,7 @@
 This module will test the main workflow of Odoo.
 It will install some main apps and will try to execute the most important actions.
 """,
-    'depends': ['web_tour', 'crm', 'sale_timesheet', 'purchase_stock', 'mrp', 'account'],
+    'depends': ['web_tour', 'crm', 'sale_project_timesheet', 'purchase_stock', 'mrp', 'account'],
     'data': [
         'views/templates.xml',
     ],
